refactor(analytics): split local state from preferences
