+++ src/services/productAnalytics/index.ts
@@ -7,4 +7,5 @@ export * from "./privacy"
 export * from "./runtime"
 export * from "./shieldBypassSummary"
 export * from "./siteEcosystem"
+export * from "./state"
 export * from "./settingsSnapshot"
