+++ tests/services/productAnalytics/runtime.test.ts
@@ -12,18 +12,25 @@ import {
   setupProductAnalyticsPreferencesChangeListener,
 } from "~/services/productAnalytics"
 
-const { captureMock, getAllAccountsMock, getPreferencesMock, preferenceMocks } =
-  vi.hoisted(() => ({
-    captureMock: vi.fn(),
-    getAllAccountsMock: vi.fn(),
-    getPreferencesMock: vi.fn(),
-    preferenceMocks: {
-      getState: vi.fn(),
-      isEnabled: vi.fn(),
-      setLastSettingsSnapshotAt: vi.fn(),
-      setLastSiteEcosystemSnapshotAt: vi.fn(),
-    },
-  }))
+const {
+  captureMock,
+  getAllAccountsMock,
+  getPreferencesMock,
+  preferenceMocks,
+  stateMocks,
+} = vi.hoisted(() => ({
+  captureMock: vi.fn(),
+  getAllAccountsMock: vi.fn(),
+  getPreferencesMock: vi.fn(),
+  preferenceMocks: {
+    isEnabled: vi.fn(),
+  },
+  stateMocks: {
+    getState: vi.fn(),
+    setLastSettingsSnapshotAt: vi.fn(),
+    setLastSiteEcosystemSnapshotAt: vi.fn(),
+  },
+}))
 
 vi.mock("~/services/productAnalytics/client", () => ({
   productAnalyticsClient: {
@@ -41,6 +48,18 @@ vi.mock("~/services/productAnalytics/preferences", async (importOriginal) => {
     productAnalyticsPreferences: {
       ...actual.productAnalyticsPreferences,
       ...preferenceMocks,
+    },
+  }
+})
+
+vi.mock("~/services/productAnalytics/state", async (importOriginal) => {
+  const actual =
+    await importOriginal<typeof import("~/services/productAnalytics/state")>()
+  return {
+    ...actual,
+    productAnalyticsState: {
+      ...actual.productAnalyticsState,
+      ...stateMocks,
     },
   }
 })
@@ -77,9 +96,9 @@ describe("product analytics runtime", () => {
     vi.setSystemTime(new Date("2026-05-12T00:00:00.000Z"))
     captureMock.mockResolvedValue(true)
     preferenceMocks.isEnabled.mockResolvedValue(true)
-    preferenceMocks.getState.mockResolvedValue({})
-    preferenceMocks.setLastSettingsSnapshotAt.mockResolvedValue(true)
-    preferenceMocks.setLastSiteEcosystemSnapshotAt.mockResolvedValue(true)
+    stateMocks.getState.mockResolvedValue({})
+    stateMocks.setLastSettingsSnapshotAt.mockResolvedValue(true)
+    stateMocks.setLastSiteEcosystemSnapshotAt.mockResolvedValue(true)
     getAllAccountsMock.mockResolvedValue([])
     getPreferencesMock.mockResolvedValue(createDefaultPreferences())
   })
@@ -134,18 +153,18 @@ describe("product analytics runtime", () => {
         webdav_configured: false,
       }),
     )
-    expect(preferenceMocks.setLastSettingsSnapshotAt).toHaveBeenCalledWith(
+    expect(stateMocks.setLastSettingsSnapshotAt).toHaveBeenCalledWith(
       Date.parse("2026-05-12T00:00:00.000Z"),
     )
     expect(sendResponse).toHaveBeenCalledWith({ success: true })
 
     vi.clearAllMocks()
     preferenceMocks.isEnabled.mockResolvedValue(true)
-    preferenceMocks.getState.mockResolvedValue({
+    stateMocks.getState.mockResolvedValue({
       lastSettingsSnapshotAt: Date.parse("2026-05-12T00:00:00.000Z"),
     })
-    preferenceMocks.setLastSiteEcosystemSnapshotAt.mockResolvedValue(true)
-    preferenceMocks.setLastSettingsSnapshotAt.mockResolvedValue(true)
+    stateMocks.setLastSiteEcosystemSnapshotAt.mockResolvedValue(true)
+    stateMocks.setLastSettingsSnapshotAt.mockResolvedValue(true)
 
     await handleProductAnalyticsMessage(
       {
@@ -157,7 +176,7 @@ describe("product analytics runtime", () => {
 
     expect(getPreferencesMock).not.toHaveBeenCalled()
     expect(captureMock).not.toHaveBeenCalled()
-    expect(preferenceMocks.setLastSettingsSnapshotAt).not.toHaveBeenCalled()
+    expect(stateMocks.setLastSettingsSnapshotAt).not.toHaveBeenCalled()
   })
 
   it("does not update settings snapshot cadence when aggregate snapshot capture returns false", async () => {
@@ -173,7 +192,7 @@ describe("product analytics runtime", () => {
     )
 
     expect(captureMock).toHaveBeenCalledTimes(1)
-    expect(preferenceMocks.setLastSettingsSnapshotAt).not.toHaveBeenCalled()
+    expect(stateMocks.setLastSettingsSnapshotAt).not.toHaveBeenCalled()
     expect(sendResponse).toHaveBeenCalledWith({ success: false })
   })
 
@@ -222,14 +241,14 @@ describe("product analytics runtime", () => {
       site_type: SITE_TYPES.NEW_API,
       account_count_bucket: "1",
     })
-    expect(preferenceMocks.setLastSiteEcosystemSnapshotAt).toHaveBeenCalledWith(
+    expect(stateMocks.setLastSiteEcosystemSnapshotAt).toHaveBeenCalledWith(
       Date.parse("2026-05-12T00:00:00.000Z"),
     )
     expect(sendResponse).toHaveBeenCalledWith({ success: true })
 
     vi.clearAllMocks()
     preferenceMocks.isEnabled.mockResolvedValue(true)
-    preferenceMocks.getState.mockResolvedValue({
+    stateMocks.getState.mockResolvedValue({
       lastSiteEcosystemSnapshotAt: Date.parse("2026-05-12T00:00:00.000Z"),
     })
 
@@ -243,9 +262,7 @@ describe("product analytics runtime", () => {
 
     expect(getAllAccountsMock).not.toHaveBeenCalled()
     expect(captureMock).not.toHaveBeenCalled()
-    expect(
-      preferenceMocks.setLastSiteEcosystemSnapshotAt,
-    ).not.toHaveBeenCalled()
+    expect(stateMocks.setLastSiteEcosystemSnapshotAt).not.toHaveBeenCalled()
   })
 
   it("does not load accounts when analytics is disabled", async () => {
@@ -285,9 +302,7 @@ describe("product analytics runtime", () => {
     )
 
     expect(captureMock).toHaveBeenCalledTimes(1)
-    expect(
-      preferenceMocks.setLastSiteEcosystemSnapshotAt,
-    ).not.toHaveBeenCalled()
+    expect(stateMocks.setLastSiteEcosystemSnapshotAt).not.toHaveBeenCalled()
     expect(sendResponse).toHaveBeenCalledWith({ success: false })
   })
 
@@ -319,9 +334,7 @@ describe("product analytics runtime", () => {
     )
 
     expect(captureMock).toHaveBeenCalledTimes(2)
-    expect(
-      preferenceMocks.setLastSiteEcosystemSnapshotAt,
-    ).not.toHaveBeenCalled()
+    expect(stateMocks.setLastSiteEcosystemSnapshotAt).not.toHaveBeenCalled()
     expect(sendResponse).toHaveBeenCalledWith({ success: false })
   })
